+++ Item/MagentoItemStep.php
@@ -2,7 +2,6 @@
 
 namespace Pim\Bundle\MagentoConnectorBundle\Item;
 
-use Pim\Bundle\MagentoConnectorBundle\Webservice\Webservice;
 use Symfony\Component\Validator\Constraints as Assert;
 use Akeneo\Bundle\BatchBundle\Item\AbstractConfigurableStepElement;
 use Pim\Bundle\MagentoConnectorBundle\Guesser\WebserviceGuesser;