+++ spec/Pim/Bundle/MagentoConnectorBundle/Cleaner/GroupCleanerSpec.php
@@ -5,7 +5,7 @@ namespace spec\Pim\Bundle\MagentoConnectorBundle\Cleaner;
 use Pim\Bundle\CatalogBundle\Entity\AttributeGroup;
 use Pim\Bundle\MagentoConnectorBundle\Entity\MagentoGroup;
 use Pim\Bundle\MagentoConnectorBundle\Guesser\WebserviceGuesser;
-use Pim\Bundle\MagentoConnectorBundle\Manager\GroupMappingManager;
+use Pim\Bundle\MagentoConnectorBundle\Manager\AttributeGroupMappingManager;
 use Pim\Bundle\MagentoConnectorBundle\Manager\MagentoGroupManager;
 use Pim\Bundle\MagentoConnectorBundle\Webservice\Webservice;
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
@@ -15,13 +15,13 @@ use Prophecy\Argument;
 class GroupCleanerSpec extends ObjectBehavior
 {
     function let(
-        WebserviceGuesser   $webserviceGuesser,
-        MagentoGroupManager $magentoGroupManager,
-        GroupMappingManager $groupMappingManager,
-        Webservice          $webservice,
-        StepExecution       $stepExecution
+        WebserviceGuesser            $webserviceGuesser,
+        MagentoGroupManager          $magentoGroupManager,
+        AttributeGroupMappingManager $attributeGroupMappingManager,
+        Webservice                   $webservice,
+        StepExecution                $stepExecution
     ) {
-        $this->beConstructedWith($webserviceGuesser, $magentoGroupManager, $groupMappingManager);
+        $this->beConstructedWith($webserviceGuesser, $magentoGroupManager, $attributeGroupMappingManager);
         $this->setStepExecution($stepExecution);
 
         $webserviceGuesser->getWebservice(Argument::cetera())->willReturn($webservice);
@@ -29,16 +29,16 @@ class GroupCleanerSpec extends ObjectBehavior
 
     function it_asks_soap_client_to_delete_groups_that_are_not_in_pim_anymore(
         $webservice,
-        GroupMappingManager $groupMappingManager,
+        AttributeGroupMappingManager $attributeGroupMappingManager,
         MagentoGroupManager $magentoGroupManager,
         AttributeGroup      $attributeGroup,
         MagentoGroup        $group
     ) {
         $group->getMagentoGroupId()->willReturn(5);
         $magentoGroupManager->getAllMagentoGroups()->willReturn(array($group));
-        $groupMappingManager->getGroupFromId(5, Argument::any())->shouldBeCalled()->WillReturn($attributeGroup);
+        $attributeGroupMappingManager->getGroupFromId(5, Argument::any())->shouldBeCalled()->WillReturn($attributeGroup);
         $attributeGroup->getCode()->shouldBeCalled();
-        $groupMappingManager->magentoGroupExists(5, Argument::any())->shouldBeCalled()->willReturn(false);
+        $attributeGroupMappingManager->magentoGroupExists(5, Argument::any())->shouldBeCalled()->willReturn(false);
 
         $webservice->removeAttributeGroupFromAttributeSet(5)->shouldBeCalled();
         $magentoGroupManager->removeMagentoGroup(5, Argument::any())->shouldBeCalled();
@@ -48,15 +48,15 @@ class GroupCleanerSpec extends ObjectBehavior
 
     function it_asks_soap_client_to_delete_groups_that_should_be_ignored(
         $webservice,
-        GroupMappingManager $groupMappingManager,
-        MagentoGroupManager $magentoGroupManager,
-        AttributeGroup      $attributeGroup,
-        MagentoGroup        $group
+        AttributeGroupMappingManager $attributeGroupMappingManager,
+        MagentoGroupManager          $magentoGroupManager,
+        AttributeGroup               $attributeGroup,
+        MagentoGroup                 $group
     ) {
         $group->getMagentoGroupId()->willReturn(4);
         $magentoGroupManager->getAllMagentoGroups()->willReturn(array($group));
-        $groupMappingManager->getGroupFromId(4, Argument::any())->shouldBeCalled()->WillReturn($attributeGroup);
-        $groupMappingManager->magentoGroupExists(4, Argument::any())->shouldBeCalled()->willReturn(false);
+        $attributeGroupMappingManager->getGroupFromId(4, Argument::any())->shouldBeCalled()->WillReturn($attributeGroup);
+        $attributeGroupMappingManager->magentoGroupExists(4, Argument::any())->shouldBeCalled()->willReturn(false);
         $attributeGroup->getCode()->shouldBeCalled()->willReturn('Default');
 
         $webservice->removeAttributeGroupFromAttributeSet(4)->shouldNotBeCalled();